requirements.txt
streamlit==1.38.0
bcrypt==4.2.0
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
psycopg2-binary>=2.9.0